Refactored HTML card
