+++ packages/kg-default-nodes/lib/nodes/html/HtmlParser.js
@@ -1,54 +1,46 @@
-export class HtmlParser {
-    constructor(NodeClass) {
-        this.NodeClass = NodeClass;
-    }
+export function parseHtmlNode(HtmlNode) {
+    return {
+        '#comment': () => {
+            return {
+                conversion(domNode) {
+                    const isCommentNode = domNode.nodeType === 8;
+                    if (isCommentNode && domNode.nodeValue.trim() === 'kg-card-begin: html') {
+                        let html = [];
+                        let nextNode = domNode.nextSibling;
 
-    get DOMConversionMap() {
-        const self = this;
-
-        return {
-            '#comment': () => {
-                return {
-                    conversion(domNode) {
-                        const isCommentNode = domNode.nodeType === 8;
-                        if (isCommentNode && domNode.nodeValue.trim() === 'kg-card-begin: html') {
-                            let html = [];
-                            let nextNode = domNode.nextSibling;
-
-                            while (nextNode && !isHtmlEndComment(nextNode)) {
-                                let currentNode = nextNode;
-                                html.push(currentNode.outerHTML);
-                                nextNode = currentNode.nextSibling;
-                                // remove nodes as we go so that they don't go through the parser
-                                currentNode.remove();
-                            }
-
-                            let payload = {html: html.join('\n').trim()};
-                            const node = new self.NodeClass(payload);
-                            return {node};
+                        while (nextNode && !isHtmlEndComment(nextNode)) {
+                            let currentNode = nextNode;
+                            html.push(currentNode.outerHTML);
+                            nextNode = currentNode.nextSibling;
+                            // remove nodes as we go so that they don't go through the parser
+                            currentNode.remove();
                         }
 
-                        return null;
+                        let payload = {html: html.join('\n').trim()};
+                        const node = new HtmlNode(payload);
+                        return {node};
+                    }
+
+                    return null;
+                },
+                priority: 0
+            };
+        },
+        table: (nodeElem) => {
+            if (nodeElem.nodeType === 1 && nodeElem.tagName === 'TABLE' && nodeElem.parentNode.tagName !== 'TABLE') {
+                return {
+                    conversion(domNode) {
+                        const payload = {html: domNode.outerHTML};
+                        const node = new HtmlNode(payload);
+                        return {node};
                     },
                     priority: 0
                 };
-            },
-            table: (nodeElem) => {
-                if (nodeElem.nodeType === 1 && nodeElem.tagName === 'TABLE' && nodeElem.parentNode.tagName !== 'TABLE') {
-                    return {
-                        conversion(domNode) {
-                            const payload = {html: domNode.outerHTML};
-                            const node = new self.NodeClass(payload);
-                            return {node};
-                        },
-                        priority: 0
-                    };
-                }
-
-                return null;
             }
-        };
-    }
+
+            return null;
+        }
+    };
 }
 
 function isHtmlEndComment(node) {
